feat: Port NodeToCode plugin from UE5.3+ to UE4.27.2
- Replace UE5-only APIs with UE4.27 equivalents:
  - PlatformAllowList -> WhitelistPlatforms in .uplugin
  - FAppStyle -> FEditorStyle
  - UAssetEditorSubsystem for editor events (works in both 4.27+)
  - FSyntaxTokenizer instead of ISyntaxTokenizer
  - GetPathName() instead of GetStructPathName()
  - ESPMode::ThreadSafe on IHttpRequest (required in 4.27)
  - TCondensedJsonPrintPolicy include for JSON serialization
  - FRegexPattern/FRegexMatcher include for Regex support
- Remove UE5-only modules: UMG, Blutility, UMGEditor
- Add UE4.27 modules: EditorStyle, ToolMenus
- Remove UE5-only features:
  - Editor Utility Widget UI system (N2CCodeEditorWidget, N2CWidgetContainer)
  - UEditorPerformanceSettings throttle override
  - PC_Double/PC_Real pin types (UE5-only)
  - K2Node_ExternalGraphInterface/PromotableOperator (UE5-only)
- Delete incompatible UE5 Content/UI assets
- Stub SMultiLineEditableText methods not in 4.27
  (DeleteSelectedText, GetCursorLocation, SelectText, GetSelection)
- Replace editor window with placeholder Slate UI for future rebuild
- Improve toolbar: labeled button with icon + dropdown for actions

+++ Source/Private/LLM/Providers/N2CLMStudioResponseParser.cpp
@@ -63,9 +63,9 @@ bool UN2CLMStudioResponseParser::ParseLLMResponse(
     const TSharedPtr<FJsonObject> StatsObject = JsonObject->GetObjectField(TEXT("stats"));
     if (StatsObject.IsValid())
     {
-        float TokensPerSecond = 0.0f;
-        float TimeToFirstToken = 0.0f;
-        float GenerationTime = 0.0f;
+        double TokensPerSecond = 0.0;
+        double TimeToFirstToken = 0.0;
+        double GenerationTime = 0.0;
         
         StatsObject->TryGetNumberField(TEXT("tokens_per_second"), TokensPerSecond);
         StatsObject->TryGetNumberField(TEXT("time_to_first_token"), TimeToFirstToken);
